requirements.txt
absl-py==2.0.0
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dlib @ https://github.com/jloh02/dlib/releases/download/v19.22/dlib-19.22.99-cp310-cp310-win_amd64.whl#sha256=f2181b6724669fb0147c9ebc326d72b94de41bced1812eff5df4b133b5d0b575
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==23.5.26
fonttools==4.45.0
idna==3.4
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
matplotlib==3.8.2
mediapipe==0.10.0
numpy==1.26.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==9.0.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.1.1
pyproject-toml==0.0.5
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
referencing==0.31.0
requests==2.31.0
rpds-py==0.12.0
schedule==1.2.1
six==1.16.0
sounddevice==0.4.6
toml==0.10.2
urllib3==2.1.0
